発話周り整理
