+++ package.json
@@ -62,15 +62,14 @@
   },
   "scripts": {
     "start": "SPEECH=off ts-node -r tsconfig-paths/register src/robot-server.ts",
-    "start:mac": "SPEECH=off ROBOT_VOICE=mac ts-node -r tsconfig-paths/register src/robot-server.ts",
+    "start:jtalk": "SPEECH=off ROBOT_DEFAULT_VOICE=open-jTalk ts-node -r tsconfig-paths/register src/robot-server.ts",
     "start:browser": "SPEECH=browser ts-node -r tsconfig-paths/register src/robot-server.ts",
     "whisper": "SPEECH=whisper WHISPER_PATH=./whisper.sh ts-node -r tsconfig-paths/register src/robot-server.ts",
-    "whisper:mac": "SPEECH=whisper WHISPER_PATH=./whisper.sh ROBOT_VOICE=mac ts-node -r tsconfig-paths/register src/robot-server.ts",
-    "reazon": "REAZON_SPEECH_HOST=host.docker.internal REAZON_SPEECH_PORT=3093 SPEECH=reazon ts-node -r tsconfig-paths/register src/robot-server.ts",
-    "reazon:mac": "REAZON_SPEECH_HOST=localhost REAZON_SPEECH_PORT=3093 SPEECH=reazon ROBOT_VOICE=mac ts-node -r tsconfig-paths/register src/robot-server.ts",
+    "reazon": "REAZON_SPEECH_HOST=localhost REAZON_SPEECH_PORT=3093 SPEECH=reazon ts-node -r tsconfig-paths/register src/robot-server.ts",
+    "reazon:docker": "REAZON_SPEECH_HOST=host.docker.internal REAZON_SPEECH_PORT=3093 SPEECH=reazon ts-node -r tsconfig-paths/register src/robot-server.ts",
     "robot": "node build/src/robot-server.js",
-    "robot:db": "SPEECH=off ROBOT_VOICE=mac ROBOT_DB=true node build/src/robot-server.js",
-    "robot:mac": "SPEECH=off ROBOT_VOICE=mac node build/src/robot-server.js",
+    "robot:db": "SPEECH=off ROBOT_DB=true node build/src/robot-server.js",
+    "robot:mac": "SPEECH=off node build/src/robot-server.js",
     "servo": "node build/src/servo-head.js",
     "servo:sudo": "sudo node build/src/servo-head.js",
     "build": "rm -rf ./build && tsc --project tsconfig.build.json && tsc-alias --verbose -p tsconfig.build.json",
