+++ src/voice.ts
@@ -1,35 +1,18 @@
-import * as EventEmitter from "events"
-import { spawn, ChildProcess } from "child_process"
-import * as path from "path"
-
+import { VoiceEmitter } from "./voice-emitter"
+import { spawn } from "child_process"
 import { config } from "./config"
 import { localhostToken } from "./accessCheck"
+import * as path from "path"
 
 const macvoice_speedrate = 210 / 100
 const axios = require("axios")
-const utils = require("./utils")
+const { basedir } = config
 
 const headers = {
   "Content-Type": "application/json",
 }
 
-const { basedir } = config
-
-class TalkEmitter extends EventEmitter {
-  playQue: any[] = []
-  playing: boolean = false
-  voice: string = "reimu"
-  speed: number = 95
-  volume: number = 80
-  robot_voice: string = null
-  languageCode?: string
-  audioEncoding?: string
-  ssmlGender?: string
-  speakingRate?: string
-  pitch?: string
-  name?: string
-  _playone?: string | ChildProcess
-
+export class Voice extends VoiceEmitter {
   constructor() {
     super()
   }
@@ -66,7 +49,7 @@ class TalkEmitter extends EventEmitter {
         playone()
         return
       }
-      console.log(text)
+      console.log(text, `languageCode:${languageCode}`, `robot_voice:${this.robot_voice}`)
       if ("host" in params) {
         // 発話の中継処理
         const p = { ...params }
@@ -86,77 +69,51 @@ class TalkEmitter extends EventEmitter {
           .catch(function (error) {
             console.error(error)
           })
-      } else if (
-        languageCode === "default" ||
-        languageCode === "normal" ||
-        //|| languageCode === 'open-jTalk'
-        !languageCode
-      ) {
+      } else if (languageCode === "default" || languageCode === "normal" || !languageCode) {
         // 通常モードの発話処理
         if (this.robot_voice === "dummy") {
           playone()
         } else if (this.robot_voice === "mac") {
           console.log("macvoice", languageCode)
-          if (languageCode === "open-jTalk") {
-            startCallback()
-            this._playone = spawn(path.join(basedir, "talk-open-jTalk-mac.sh"), [
-              voice === "reimu" ? "mei_normal" : voice,
-              `${text}`,
-            ])
-            this._playone.on("close", function (code) {
-              playone()
-            })
-          } else {
-            startCallback()
-            this._playone = spawn(path.join(basedir, "talk-mac.sh"), [
-              `-r`,
-              speed * macvoice_speedrate,
-              text,
-              voice === "reimu" ? "" : voice || "",
-            ])
-            this._playone.on("close", function (code) {
-              playone()
-            })
-          }
+          startCallback()
+          this._playone = spawn(path.join(basedir, "talk-mac.sh"), [
+            `-r`,
+            speed * macvoice_speedrate,
+            text,
+            voice === "reimu" ? "" : voice || "",
+          ])
+          this._playone.on("close", function (code) {
+            playone()
+          })
         } else {
-          if (languageCode === "open-jTalk") {
+          if (voice == "marisa") {
             startCallback()
-            this._playone = spawn(path.join(basedir, "talk-open-jTalk-raspi.sh"), [
-              voice === "reimu" ? "mei_normal" : voice,
-              `${text}`,
+            this._playone = spawn(path.join(basedir, "talk-f2.sh"), [
+              `-s`,
+              speed,
+              `-g`,
+              volume,
+              `　${text}`,
             ])
             this._playone.on("close", function (code) {
               playone()
             })
           } else {
-            if (voice == "marisa") {
-              startCallback()
-              this._playone = spawn(path.join(basedir, "talk-f2.sh"), [
-                `-s`,
-                speed,
-                `-g`,
-                volume,
-                `　${text}`,
-              ])
-              this._playone.on("close", function (code) {
-                playone()
-              })
-            } else {
-              startCallback()
-              this._playone = spawn(path.join(basedir, "talk-f1.sh"), [
-                `-s`,
-                speed,
-                `-g`,
-                volume,
-                `　${text}`,
-              ])
-              this._playone.on("close", function (code) {
-                playone()
-              })
-            }
+            startCallback()
+            this._playone = spawn(path.join(basedir, "talk-f1.sh"), [
+              `-s`,
+              speed,
+              `-g`,
+              volume,
+              `　${text}`,
+            ])
+            this._playone.on("close", function (code) {
+              playone()
+            })
           }
         }
       } else if (languageCode) {
+        // open-jtalk / google text-to-speech / aws poly の場合
         // 言語指定の発話処理(google-router.jsへリクエスト)
         const params = {
           text,
@@ -216,112 +173,10 @@ class TalkEmitter extends EventEmitter {
     }
     playone()
   }
-
-  _playAsync(speech: string, params, callback) {
-    return new Promise<void>((resolve) => {
-      let doneStart = false
-      this._say(
-        speech,
-        params,
-        () => {
-          resolve()
-        },
-        () => {
-          if (!doneStart) {
-            if (callback) callback("talk")
-          }
-          doneStart = true
-        }
-      )
-    })
-  }
-
-  play(
-    sentence,
-    params: {
-      voice?: string
-      speed?: number
-      volume?: number
-      languageCode?: string
-      audioEncoding?: string
-      ssmlGender?: string
-      speakingRate?: string
-      pitch?: string
-      name?: string
-    } = {},
-    callback = null
-  ) {
-    //デフォルトパラメータの設定
-    {
-      if (typeof params.voice === "undefined") params.voice = this.voice
-      if (typeof params.speed === "undefined") params.speed = this.speed
-      if (typeof params.volume === "undefined") params.volume = this.volume
-      if (typeof params.languageCode === "undefined") params.languageCode = this.languageCode
-      if (typeof params.audioEncoding === "undefined") params.audioEncoding = this.audioEncoding
-      if (typeof params.ssmlGender === "undefined") params.ssmlGender = this.ssmlGender
-      if (typeof params.speakingRate === "undefined") params.speakingRate = this.speakingRate
-      if (typeof params.pitch === "undefined") params.pitch = this.pitch
-      if (typeof params.name === "undefined") params.name = this.name
-    }
-    this.emit("talk")
-    if (!this.playing) {
-      this.playing = true
-      const _play = (sentence) => {
-        this._playAsync(sentence, params, callback).then(() => {
-          if (this.playQue.length > 0 && this.playing !== false) {
-            const sentence = this.playQue.shift()
-            _play(sentence)
-          } else {
-            this.playQue = []
-            this.playing = false
-            this.emit("idle")
-            if (callback) callback("idle")
-          }
-        })
-      }
-      _play(sentence)
-    } else {
-      this.playQue.push(sentence)
-    }
-  }
-
-  stop(callback = null) {
-    this.playing = false
-    if (this._playone) {
-      if (typeof this._playone === "string") {
-        axios
-          .post(
-            `${this._playone}/text-to-speech`,
-            {
-              action: "stop",
-              localhostToken: localhostToken(),
-            },
-            { headers }
-          )
-          .then(function (response) {
-            if (callback) callback()
-          })
-          .catch(function (error) {
-            console.error(error)
-          })
-        this._playone = null
-        return
-      } else {
-        utils.kill(this._playone.pid, "SIGTERM", function () {})
-      }
-    }
-    this._playone = null
-    if (callback) callback()
-  }
-
-  flush() {
-    this.playing = false
-    this.playQue = []
-  }
 }
 
 export function Talk() {
-  return new TalkEmitter()
+  return new Voice()
 }
 
 function main() {
