+++ talk-open-jTalk.sh
@@ -1,11 +1,13 @@
 #!/bin/sh
 #
-# OpenJTalk(RaspberryPi用) を使って発話する
+# OpenJTalk を使って発話する
 #
 
+cd `dirname $0`
 TMP="$3"
+USER_NAME=`whoami`
 VOICE_NAME="$1"
-ALL_VOICE_PATH=`find ~/ -name *"$VOICE_NAME"*.htsvoice`
+ALL_VOICE_PATH=`find ./modules/open_jtalk/ -name *"$VOICE_NAME"*.htsvoice`
 # echo $ALL_VOICE_PATH
 for var in $ALL_VOICE_PATH
 do
@@ -15,5 +17,5 @@ do
 done
 echo "$2" | open_jtalk \
 -m $VOICE_PATH \
--x "/var/lib/mecab/dic/open-jtalk/naist-jdic/" \
+-x ./modules/open_jtalk/open_jtalk_dic_utf_8-1.11 \
 -ow $TMP
